refactor(lib): move sanitizeFilename out of zoho-sign provider into shared lib
