+++ src/shared/services/providers/zoho-sign/lib/documents/assemble-envelope.ts
@@ -1,9 +1,9 @@
 import type { AttachFile } from '../../client'
 import type { EnvelopeDocument, ProposalContext } from './types'
 import type { EnvelopeDocumentId } from '@/shared/constants/enums'
+import { sanitizeFilename } from '@/shared/lib/sanitize-filename'
 import { zohoSignClient } from '../../client'
 import { SOW_INLINE_MAX_CHARS } from '../../constants'
-import { sanitizeFilename } from '../sanitize-filename'
 import { evaluateDocuments } from './evaluate'
 import { ENVELOPE_DOCUMENTS } from './registry'
