+++ src/shared/services/zoho-sync.service.ts
@@ -1,9 +1,9 @@
 import type { ProposalContext } from './providers/zoho-sign/lib/documents/types'
 import type { ZohoContractStatus, ZohoRequestStatus } from './providers/zoho-sign/types'
+import { sanitizeFilename } from '@/shared/lib/sanitize-filename'
 import { zohoSignClient } from './providers/zoho-sign/client'
 import { dedupeSignerStatuses } from './providers/zoho-sign/lib/dedupe-signer-statuses'
 import { assembleEnvelope } from './providers/zoho-sign/lib/documents/assemble-envelope'
-import { sanitizeFilename } from './providers/zoho-sign/lib/sanitize-filename'
 
 // ---------------------------------------------------------------------------
 // Result types
